requirements.txt
ms-swift[rlhf]
math-verify
vllm
torch
deepspeed
wandb
